// src/http/generated/models/CyberAssetNotificationsTypes.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * MonoWatch BackBone
 * MonoWatch Security Asset Management
 * OpenAPI spec version: 1.0.0
 */

export type CyberAssetNotificationsTypes = typeof CyberAssetNotificationsTypes[keyof typeof CyberAssetNotificationsTypes];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CyberAssetNotificationsTypes = {
  error: 'error',
  info: 'info',
  warning: 'warning',
  config: 'config',
} as const;