// src/http/generated/models/EachCyberAssetAvailabilityStatus.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * MonoWatch BackBone
 * MonoWatch Security Asset Management
 * OpenAPI spec version: 1.0.0
 */

export type EachCyberAssetAvailabilityStatus = typeof EachCyberAssetAvailabilityStatus[keyof typeof EachCyberAssetAvailabilityStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EachCyberAssetAvailabilityStatus = {
  online: 'online',
  offline: 'offline',
  unmanaged: 'unmanaged',
} as const;